PARQUET-288: Add dictionary support to other converters.
This doesn't add dictionary support for long, float, and double because
it would be unsafe. Those converters can be used for columns that have a
different underlying type and will automatically promote the type. For
example, a long converter can be used with an int column. In order for
this to be safe, the converter needs to know the file's type for the
column, which isn't currently passed.