Add new test encodings.
This adds two new encoders that implement a variable-width RLE encoding.
The new encoding is identical to the existing RLE except that after the
bit-packed header byte there is another byte that encodes the width of
the bit packed run.

The variable-width RLE encoder is used by writer variants that zig-zag
encode values (move the sign bit to lsb) to support negative numbers in
the RLE encoding. Other writers add integer delta encoding, decompose
floating points to store the exponent and significand separately, and
xor the significand (similar to delta-encoding for floats).

This relies on format changes:
  https://github.com/rdblue/parquet-format/tree/zig-zag-encodings